refactor(debugger): address PR review comments
- batch realtime store entry points (applyRealtimeSpans / applyTraceUpdates)
- seed projectId at store creation instead of effect-syncing the URL param
- formatGap -> formatDuration, moved to lib/utils with the other time helpers
- drop narration comments flagged in review
