+++ frontend/lib/utils.ts
@@ -191,6 +191,28 @@ function innerFormatTimestamp(date: Date, format?: string): string {
   return `${dateStr}, ${timeStr}`;
 }
 
+/** Format a duration in ms as a short human-readable string (e.g. "3m 12s").
+ *  Returns null for zero/negative/invalid durations. */
+export function formatDuration(ms: number | undefined): string | null {
+  if (ms === undefined || !Number.isFinite(ms) || ms <= 0) return null;
+  const seconds = Math.floor(ms / 1000);
+  if (seconds < 1) return "<1s";
+  if (seconds < 60) return `${seconds}s`;
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) {
+    const s = seconds % 60;
+    return s === 0 ? `${minutes}m` : `${minutes}m ${s}s`;
+  }
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) {
+    const m = minutes % 60;
+    return m === 0 ? `${hours}h` : `${hours}h ${m}m`;
+  }
+  const days = Math.floor(hours / 24);
+  const h = hours % 24;
+  return h === 0 ? `${days}d` : `${days}d ${h}h`;
+}
+
 export function deep<T>(value: T): T {
   if (typeof value !== "object" || value === null) {
     return value;
